requirements.txt
Flask
SQLAlchemy
python-dotenv
langchain_openai
langchain_community
langchain_core
faiss-cpu
watchdog==2.1.6